feat: testes e estruturacao inicial do recurso de produto

+++ grpc-spring-study/src/main/java/com/fredsonchaves07/application/dto/ProductOutputDTO.java
@@ -1,4 +1,4 @@
 package com.fredsonchaves07.application.dto;
 
-public record ProductOutputDTO(Integer id, String name, Double price, Integer quantityInStock) {
+public record ProductOutputDTO(Long id, String name, Double price, Integer quantityInStock) {
 }
